requirements.txt
beautifulsoup4==4.12.3
python-telegram-bot==21.4
requests==2.31.0
seleniumbase==4.29.9
webdriver-manager==4.0.2
python-telegram-bot[job-queue]